[RW-6670][risk=low] Fix DataSetController access controls (#4918)

+++ api/src/main/java/org/pmiops/workbench/dataset/DataSetServiceImpl.java
@@ -247,14 +247,20 @@ public class DataSetServiceImpl implements DataSetService, GaugeDataCollector {
   }
 
   @Override
-  public DataSet updateDataSet(DataSetRequest request, Long dataSetId) {
-    DbDataset dbDataSet = dataSetDao.findOne(dataSetId);
+  public DataSet updateDataSet(long workspaceId, long dataSetId, DataSetRequest request) {
+    Optional<DbDataset> dbDataSet =
+        dataSetDao.findByDataSetIdAndWorkspaceId(dataSetId, workspaceId);
+
+    if (!dbDataSet.isPresent()) {
+      throw new NotFoundException(
+          "No DataSet found for dataSetId " + dataSetId + " and workspaceId " + workspaceId);
+    }
 
     int version = Etags.toVersion(request.getEtag());
-    if (dbDataSet.getVersion() != version) {
+    if (dbDataSet.get().getVersion() != version) {
       throw new ConflictException("Attempted to modify outdated data set version");
     }
-    DbDataset dbMappingConvert = dataSetMapper.dataSetRequestToDb(request, dbDataSet, clock);
+    DbDataset dbMappingConvert = dataSetMapper.dataSetRequestToDb(request, dbDataSet.get(), clock);
     return saveDataSet(dbMappingConvert);
   }
 
@@ -812,43 +818,61 @@ public class DataSetServiceImpl implements DataSetService, GaugeDataCollector {
     return cohortDao.findAllByCohortIdIn(dataSetDao.findOne(dataSet.getDataSetId()).getCohortIds());
   }
 
-  public List<DataSet> getDataSets(ResourceType resourceType, long resourceId) {
-    return getDbDataSets(resourceType, resourceId).stream()
+  public List<DataSet> getDataSets(long workspaceId, ResourceType resourceType, long resourceId) {
+    return getDbDataSets(workspaceId, resourceType, resourceId).stream()
         .map(dataSetMapper::dbModelToClient)
         .collect(Collectors.toList());
   }
 
-  public List<DbDataset> getDbDataSets(ResourceType resourceType, long resourceId) {
+  // TODO(calbach): Remove direct testing of this - cover via public interface.
+  @VisibleForTesting
+  public List<DbDataset> getDbDataSets(
+      long workspaceId, ResourceType resourceType, long resourceId) {
     List<DbDataset> dbDataSets = new ArrayList<>();
     switch (resourceType) {
       case COHORT:
-        dbDataSets = dataSetDao.findDataSetsByCohortIds(resourceId);
+        DbCohort dbCohort = cohortDao.findOne(resourceId);
+        if (dbCohort.getWorkspaceId() != workspaceId) {
+          throw new NotFoundException("Resource does not belong to specified workspace");
+        }
+        dbDataSets = dataSetDao.findDataSetsByCohortIdsAndWorkspaceId(resourceId, workspaceId);
         break;
       case CONCEPT_SET:
-        dbDataSets = dataSetDao.findDataSetsByConceptSetIds(resourceId);
+        DbConceptSet dbConceptSet = conceptSetDao.findOne(resourceId);
+        if (dbConceptSet.getWorkspaceId() != workspaceId) {
+          throw new NotFoundException("Resource does not belong to specified workspace");
+        }
+        dbDataSets = dataSetDao.findDataSetsByConceptSetIdsAndWorkspaceId(resourceId, workspaceId);
         break;
     }
     return dbDataSets;
   }
 
   @Override
-  public void deleteDataSet(Long dataSetId) {
+  public void deleteDataSet(long workspaceId, long dataSetId) {
+    Optional<DbDataset> dbDataset = this.getDbDataSet(workspaceId, dataSetId);
+    if (!dbDataset.isPresent()) {
+      throw new NotFoundException(
+          "No DataSet found for dataSetId " + dataSetId + " and workspaceId " + workspaceId);
+    }
     dataSetDao.delete(dataSetId);
   }
 
   @Override
-  public Optional<DataSet> getDataSet(Long dataSetId) {
-    return getDbDataSet(dataSetId).map(dataSetMapper::dbModelToClient);
+  public Optional<DataSet> getDataSet(long workspaceId, long dataSetId) {
+    return dataSetDao
+        .findByDataSetIdAndWorkspaceId(dataSetId, workspaceId)
+        .map(dataSetMapper::dbModelToClient);
   }
 
   @Override
-  public Optional<DbDataset> getDbDataSet(Long dataSetId) {
-    return Optional.ofNullable(dataSetDao.findOne(dataSetId));
+  public Optional<DbDataset> getDbDataSet(long workspaceId, long dataSetId) {
+    return dataSetDao.findByDataSetIdAndWorkspaceId(dataSetId, workspaceId);
   }
 
   @Override
-  public void markDirty(ResourceType resourceType, long resourceId) {
-    List<DbDataset> dbDataSetList = getDbDataSets(resourceType, resourceId);
+  public void markDirty(long workspaceId, ResourceType resourceType, long resourceId) {
+    List<DbDataset> dbDataSetList = getDbDataSets(workspaceId, resourceType, resourceId);
     dbDataSetList.forEach(dataSet -> dataSet.setInvalid(true));
     try {
       dataSetDao.save(dbDataSetList);
